% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{create_pdf_proj}
\alias{create_pdf_proj}
\title{Create PDF formated}
\usage{
create_pdf_proj(file, format = "both", output_dir = "./pdfs")
}
\arguments{
\item{file}{A path of RMarkdown file with SEDAP Research Project Template;}

\item{format}{Format of research project pdf, can be "book", "online" or "both"; \strong{\preformatted{Default: "both"
}

}}

\item{output_dir}{A output directory of pdfs files;}
}
\value{
PDF file in output_dir.
}
\description{
A function to make a pdf files by RMarkdown template of SEDAP research project.
}
